# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/bussdev/lead": {
      "post": {
        "operationId": "BussdevController_createLead",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLeadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new business lead",
        "tags": [
          "bussdev"
        ]
      }
    },
    "/bussdev/lead/{id}/advance": {
      "patch": {
        "operationId": "BussdevController_advanceLead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvanceLeadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Advance a lead to the next stage",
        "tags": [
          "bussdev"
        ]
      }
    },
    "/bussdev/dashboard": {
      "get": {
        "operationId": "BussdevController_getDashboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "bussdev"
        ]
      }
    },
    "/bussdev/analytics/funnel": {
      "get": {
        "operationId": "BussdevController_getFunnelAnalytics",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "bussdev"
        ]
      }
    },
    "/bussdev/analytics/pipeline-granular": {
      "get": {
        "operationId": "BussdevController_getGranularPipelineTable",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "bussdev"
        ]
      }
    },
    "/bussdev/analytics/staff-performance": {
      "get": {
        "operationId": "BussdevController_getBDPerformance",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "bussdev"
        ]
      }
    },
    "/bussdev/analytics/lost-churn": {
      "get": {
        "operationId": "BussdevController_getLostChurnTable",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "bussdev"
        ]
      }
    },
    "/bussdev/analytics/{group}": {
      "get": {
        "operationId": "BussdevController_getPageAnalytics",
        "parameters": [
          {
            "name": "group",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "bussdev"
        ]
      }
    },
    "/bussdev/leads": {
      "get": {
        "operationId": "BussdevController_getLeads",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all leads",
        "tags": [
          "bussdev"
        ]
      }
    },
    "/bussdev/leads/stuck": {
      "get": {
        "operationId": "BussdevController_getStuckLeads",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "bussdev"
        ]
      }
    },
    "/bussdev/leads/group/{group}": {
      "get": {
        "operationId": "BussdevController_getLeadsByGroup",
        "parameters": [
          {
            "name": "group",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "bussdev"
        ]
      }
    },
    "/bussdev/staffs": {
      "get": {
        "operationId": "BussdevController_getStaffs",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "bussdev"
        ]
      }
    },
    "/bussdev/samples": {
      "get": {
        "operationId": "BussdevController_getClientSamples",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "bussdev"
        ]
      }
    },
    "/bussdev/sample/{id}/ship": {
      "patch": {
        "operationId": "BussdevController_shipSample",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "bussdev"
        ]
      }
    },
    "/bussdev/sample/{id}/feedback": {
      "patch": {
        "operationId": "BussdevController_submitFeedback",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "bussdev"
        ]
      }
    },
    "/bussdev/lead/{id}/activity": {
      "post": {
        "operationId": "BussdevController_logActivity",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Log a text interaction (Chat/Call/Meeting)",
        "tags": [
          "bussdev"
        ]
      }
    },
    "/bussdev/lead/{id}/activity-stream": {
      "get": {
        "operationId": "BussdevController_getActivityStream",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "bussdev"
        ]
      }
    },
    "/bussdev/lead/{id}/balance": {
      "get": {
        "operationId": "BussdevController_getLeadBalance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "bussdev"
        ]
      }
    },
    "/bussdev/guest/{id}/convert": {
      "post": {
        "operationId": "BussdevController_convertGuestToLead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "bussdev"
        ]
      }
    },
    "/bussdev/sales-order/{id}/status": {
      "patch": {
        "operationId": "BussdevController_updateSoStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "bussdev"
        ]
      }
    },
    "/bussdev/lead/{id}/override": {
      "patch": {
        "operationId": "BussdevController_emergencyOverride",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "bussdev"
        ]
      }
    },
    "/bussdev/debug/leads": {
      "get": {
        "operationId": "BussdevController_debugLeads",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "bussdev"
        ]
      }
    },
    "/bussdev/retention-engine/{id}/trigger": {
      "post": {
        "operationId": "BussdevController_triggerRetention",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "bussdev"
        ]
      }
    },
...
  },
  "info": {
    "title": "NexERP API",
    "description": "The ultimate manufacturing ERP system API documentation",
    "version": "4.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateSampleRequestDto": {
        "type": "object",
        "properties": {
          "productName": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "targetPrice": {
            "type": "number"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "productName"
        ]
      },
      "CreateLeadDto": {
        "type": "object",
        "properties": {
          "clientName": {
            "type": "string"
          },
          "brandName": {
            "type": "string"
          },
          "contactInfo": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "productInterest": {
            "type": "string"
          },
          "estimatedValue": {
            "type": "number"
          },
          "picId": {
            "type": "string",
            "format": "uuid"
          },
          "isRepeatOrder": {
            "type": "boolean"
          },
          "hkiMode": {
            "type": "object"
          },
          "paymentType": {
            "type": "object"
          },
          "category": {
            "type": "object"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "province": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "addressDetail": {
            "type": "string"
          },
          "launchingPlan": {
            "type": "string"
          },
          "targetMarket": {
            "type": "string"
          },
          "contactChannel": {
            "type": "string"
          },
          "packagingSuggestion": {
            "type": "string"
          },
          "designSuggestion": {
            "type": "string"
          },
          "valueSuggestion": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "unitPrice": {
            "type": "number"
          },
          "notes": {
            "type": "string"
          },
          "moq": {
            "type": "number"
          },
          "planOmset": {
            "type": "number"
          },
          "logoRevision": {
            "type": "number"
          },
          "hkiProgress": {
            "type": "string"
          },
          "sampleRequests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateSampleRequestDto"
            }
          }
        },
        "required": [
          "clientName",
          "contactInfo",
          "source",
          "productInterest",
          "estimatedValue"
        ]
      },
      "AdvanceLeadDto": {
        "type": "object",
        "properties": {
          "action": {
            "type": "object"
          },
          "newStatus": {
            "type": "object"
          },
          "lostReason": {
            "type": "object"
          },
          "paymentType": {
            "type": "object"
          },
          "notes": {
            "type": "string"
          },
          "isRepeatOrder": {
            "type": "boolean"
          },
          "pnfFileUrl": {
            "type": "string"
          },
          "spkFileUrl": {
            "type": "string"
          },
          "clientExpectations": {
            "type": "string"
          },
          "paymentProofUrl": {
            "type": "string"
          },
          "downPaymentAmount": {
            "type": "number"
          },
          "productConcept": {
            "type": "string"
          },
          "targetPrice": {
            "type": "number"
          },
          "loggedBy": {
            "type": "string"
          },
          "productCategory": {
            "type": "string"
          },
          "estimatedMoq": {
            "type": "number"
          },
          "planOmset": {
            "type": "number"
          },
          "packagingSuggestion": {
            "type": "string"
          },
          "designSuggestion": {
            "type": "string"
          },
          "valueSuggestion": {
            "type": "string"
          }
        },
        "required": [
          "action",
          "newStatus",
          "loggedBy"
        ]
      },
...
    }
  }
}